♻️ Refactor: 일반적인 container구조로 folderPage적용
